+++ example/lib/model/item.dart
@@ -1,19 +1,22 @@
 import 'package:flamingo/flamingo.dart';
 import 'package:flamingo_example/model/setting.dart';
 
-class User extends Document<User> {
-  User({
+class Item extends Document<Item> {
+  Item({
     String id,
     DocumentSnapshot snapshot,
     Map<String, dynamic> values,
-  }): super(id: id, snapshot: snapshot, values: values) {
-    setting = Collection(this, 'setting');
-  }
+  }): super(id: id, snapshot: snapshot, values: values);
 
   String name;
-  Collection<Setting> setting;
 
-  /// For save data
+  /// Custom field key of createdAt and updatedAt.
+  @override
+  String get createdFieldValueKey => 'createdDate';
+
+  @override
+  String get updatedFieldValueKey => 'updatedDate';
+
   @override
   Map<String, dynamic> toData() {
     final data = <String, dynamic>{};
@@ -21,20 +24,13 @@ class User extends Document<User> {
     return data;
   }
 
-  /// For load data
   @override
   void fromData(Map<String, dynamic> data) {
     name = valueFromKey<String>(data, 'name');
   }
 
-  /// For completed create, update, delete.
-  @override
-  void onCompleted(ExecuteType executeType) {
-    print('onCompleted $executeType');
-  }
-
   void log() {
-    print('User $id ${reference.path} $name $createdFieldValueKey $updatedFieldValueKey '
+    print('Item $id ${reference.path} $name $createdFieldValueKey $updatedFieldValueKey '
         '${createdAt?.toDate()} '
         '${updatedAt?.toDate()}');
   }