+++ lib/constants.dart
@@ -1,0 +1,3 @@
+const documentCreatedAtKey = 'createdAt';
+const documentUpdatedAtKey = 'updatedAt';
+const mimeTypeApplicationOctetStream = 'application/octet-stream';